+++ out/getHtml.js
@@ -24,8 +24,8 @@ function createEditorHtml(webview, context, config, initialVars) {
     const _getResourcePath = getResourcePath.bind(undefined, webview, context);
     let handsontableCss = _getResourcePath('thirdParty/handsontable/handsontable.min.css');
     // let handsontableCss = _getResourcePath('thirdParty/handsontable/handsontable.css')
-    // let handsontableJs = _getResourcePath('thirdParty/handsontable/handsontable.min.js')
-    let handsontableJs = _getResourcePath('thirdParty/handsontable/handsontable.js');
+    let handsontableJs = _getResourcePath('thirdParty/handsontable/handsontable.min.js');
+    // let handsontableJs = _getResourcePath('thirdParty/handsontable/handsontable.js')
     let papaparseJs = _getResourcePath('thirdParty/papaparse/papaparse.min.js');
     // let papaparseJs = _getResourcePath('thirdParty/papaparse/papaparse.js')
     let regressionJS = _getResourcePath('thirdParty/regression/regression.min.umd.js');