+++ csvEditorHtml/util.ts
@@ -1070,7 +1070,6 @@ function setupAndApplyInitialConfigPart1(initialConfig: EditCsvConfig | undefine
 		escapeChar: initialConfig.writeOption_escapeChar,
 		quoteChar: initialConfig.writeOption_quoteChar,
 		quoteAllFields: initialConfig.quoteAllFields,
-		retainQuoteInformation: initialConfig.retainQuoteInformation,
 		quoteEmptyOrNullFields: initialConfig.quoteEmptyOrNullFields === 'true' ? true : false,
 	})